requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
loguru==0.6.0
PyYAML==6.0
requests==2.27.1
urllib3==1.26.9
